Added Database Generator And Integration Tests
Made it far easier to generate and test a database.

It appears that the database algorithm is too slow as it tries to find every possible match for every star pair.
This requires a long comparison to verify that it is a correct star.
i.e. if the # pairs of a triangle found are 400, 300, 500. It will have to run every combination to find a matching triangle.

A potential fix will be to have multiple databases of different "regions" of the sky