Added Star Triangles Implementation (Porely)
The reason why star triangles were not generated was because I forgot to put it in the code.
Now it is in tracking_mode::constellation::find().

It does not take priority of the triangle and will use the last found which is not as accurate as using a weighting system.